Tutorial Hide Prompt + Slam Hit Interval Cap + Projectile Speed Cap
Player is now prompted to move when under the tutorial prompt area. Slam prefabs now only hit for the first 0.5s of their lifespan, this was frustrating players as they would move into the slam late and still be hit by it. Capped projectile speed at 14 (default is 7) as some projectiles were becoming insanely fast due to the difficulty, this is confusing as the logical maximum value should be only 14 also.